requirements.txt
python-nmap==0.7.1
requests==2.31.0